Add -gomod flag to speed up and harmonize vendoring
Go's module support is a welcome improvement for the ecosystem,
however it is currently unfit for some larger projects that
had git tags prior to go mod's arrival, or projects that depend on
one another and have their dependencies change frequently.

As much as I'd like to help go mod evolve to address those cases,
today vndr just works™, but is too slow compared to go mod.

This patch is backwards compatible but if -gomod is specified
or the VNDR_GOMOD environment variable is set to a non-zero value
then it ignores most of the codepath and simply generates a dumb
temporary go.mod file that abuses the replace directive, before
running go mod vendor and removing go.mod and go.sum.

The -verbose flag adds -v to the go mod vendor command as well as
outputs the content of go.mod before removing it.

Signed-off-by: Tibor Vass <tibor@docker.com>